apps/backend/pyproject.toml
[project]
name = "examnova-backend"
version = "0.1.0"
description = "Backend skeleton for ExamNova"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.27.2",
  "pillow>=10.0.0",
  "python-multipart>=0.0.9",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.0"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["data*"]
